Add ctucanfd_ip_core - CAN with Flexible Data-rate IP Core developed at Department of Measurement of FEE CTU
CTU CAN FD is soft IP-Core written in VHDL which supports ISO and NON-ISO versions of CAN FD protocol.
CTU CAN FD is NOT a prototype. RTL is compliant with ISO 1198-1:2015. It is tested by ISO 16845-1 2016 sequence in regression run every day.
RTL is implemented with VHDL-93, synthesizable with most FPGA and ASIC flows. TB is implemented with VHDL 2008.